Use Int64 for Limit values
- this corresponds with the way limits are stored in both the `data` package and the `mongodb` package.

+++ txn/domain-blocks.go
@@ -15,7 +15,7 @@ type GetDomainBlocks struct {
 	MaxID         string `query:"max_id"`
 	SinceID       string `query:"since_id"`
 	MinID         string `query:"min_id"`
-	Limit         int    `query:"limit"`
+	Limit         int64  `query:"limit"`
 }
 
 // QueryPage implements the QueryPager interface, returning
