+++ txn/followed-tags.go
@@ -14,7 +14,7 @@ type GetFollowedTags struct {
 	MaxID   string `query:"max_id"`
 	SinceID string `query:"since_id"`
 	MinID   string `query:"min_id"`
-	Limit   int    `query:"limit"`
+	Limit   int64  `query:"limit"`
 }
 
 // QueryPage implements the QueryPager interface, returning
