+++ txn/lists.go
@@ -57,7 +57,7 @@ type GetList_Accounts struct {
 	MaxID   string `query:"max_id"`
 	SinceID string `query:"since_id"`
 	MinID   string `query:"min_id"`
-	Limit   int    `query:"limit"`
+	Limit   int64  `query:"limit"`
 }
 
 // QueryPage implements the QueryPager interface, returning
