+++ txn/search.go
@@ -20,7 +20,7 @@ type GetSearch struct {
 	MaxID             string `query:"max_id"`
 	MinID             string `query:"min_id"`
 	SinceID           string `query:"since_id"`
-	Limit             int    `query:"limit"`
+	Limit             int64  `query:"limit"`
 	Offset            int    `query:"offset"`
 }
